counters-mini/requirements.txt
python-dotenv~=1.0.0
jsonschema~=4.19.1
tekore~=5.0.1
PyGithub~=1.59.1
typing_extensions~=4.8.0
